Split Push provider into controller/provider
