+++ lib/Alchemy/Phrasea/Application.php
@@ -12,7 +12,6 @@
 namespace Alchemy\Phrasea;
 
 use Alchemy\Geonames\GeonamesServiceProvider;
-use Alchemy\Phrasea\ControllerProvider\Prod\Push;
 use Alchemy\Phrasea\ControllerProvider\Prod\Query;
 use Alchemy\Phrasea\ControllerProvider\Prod\Records;
 use Alchemy\Phrasea\ControllerProvider\Prod\Root as Prod;
@@ -315,6 +314,7 @@ class Application extends SilexApplication
             'Alchemy\Phrasea\ControllerProvider\Prod\Order' => [],
             'Alchemy\Phrasea\ControllerProvider\Prod\Printer' => [],
             'Alchemy\Phrasea\ControllerProvider\Prod\Property' => [],
+            'Alchemy\Phrasea\ControllerProvider\Prod\Push' => [],
             'Alchemy\Phrasea\ControllerProvider\Datafiles' => [],
             'Alchemy\Phrasea\ControllerProvider\Lightbox' => [],
             'Alchemy\Phrasea\ControllerProvider\MediaAccessor' => [],
@@ -630,7 +630,6 @@ class Application extends SilexApplication
         $this->mount('/prod/WorkZone', new WorkZone());
         $this->mount('/prod/lists', new UsrLists());
         $this->mount('/prod/records/', new Records());
-        $this->mount('/prod/push/', new Push());
         $this->mount('/prod/share/', new Share());
         $this->mount('/prod/TOU/', new TOU());
         $this->mount('/prod/tooltip', new Tooltip());
@@ -679,6 +678,7 @@ class Application extends SilexApplication
             '/prod/lazaret/'               => 'Alchemy\Phrasea\ControllerProvider\Prod\Lazaret',
             '/prod/order/'                 => 'Alchemy\Phrasea\ControllerProvider\Prod\Order',
             '/prod/printer/'               => 'Alchemy\Phrasea\ControllerProvider\Prod\Printer',
+            '/prod/push/'                  => 'Alchemy\Phrasea\ControllerProvider\Prod\Push',
             '/prod/records/edit'           => 'Alchemy\Phrasea\ControllerProvider\Prod\Edit',
             '/prod/records/movecollection' => 'Alchemy\Phrasea\ControllerProvider\Prod\MoveCollection',
             '/prod/records/property'       => 'Alchemy\Phrasea\ControllerProvider\Prod\Property',
